分离函数
